+++ examples/batch_mode/01-full_basic.py
@@ -39,13 +39,14 @@ def plot(plot_dir):
     y = -0.0
     dx = 5.
     nx = 2400
+    fs.settings.enable_multithreading = False
     
     calc_name="mandelbrot"
     colormap = fscolors.cmap_register["classic"]
 
     # Run the calculation
     f = fsm.Mandelbrot(plot_dir)
-    f.zoom(x=x, y=y, dx=dx, nx=nx, xy_ratio=1.0,
+    f.zoom(x=x, y=y, dx=dx, nx=nx, xy_ratio=4.0,
            theta_deg=0., projection="cartesian", antialiasing=False)
     f.base_calc(
         calc_name=calc_name,